requirements/basse.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
biopython==1.69
Django==4.0.1
django-blastplus==3.0.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2